LSTM E2E working

+++ common/utils/utils.py
@@ -63,7 +63,7 @@ def print_progress_bar(iteration, total, prefix='', suffix='', decimals=1, lengt
     percent = ("{0:." + str(decimals) + "f}").format(100 * (iteration / float(total)))
     filled_length = int(length * iteration // total)
     bar = fill * filled_length + '-' * (length - filled_length)
-    print('\r%s |%s| %s%% %s' % (prefix, bar, percent, suffix), end = '\r')
+    print('\r%s |%s| %s%% %s' % (prefix, bar, percent, suffix), end='\r')
     # Print New Line on Complete
     if iteration == total:
         print()
@@ -109,13 +109,10 @@ def word_dropout(dropout_prob: float) -> bool:
 
 def get_onehot_np(y: np.array, no_classes: int):
     # np.eye won't work for float which we use to get one hot
-    y_onehot_np = np.eye(no_classes, dtype=np.float32)[y]
+    y_onehot_np = np.eye(no_classes, dtype=np.int32)[y]
     y_onehot_np = np.squeeze(y_onehot_np, axis=0)
     return y_onehot_np
 
 
 def count_parameters(model):
     return sum(p.numel() for p in model.parameters() if p.requires_grad)
-
-
-
